Move live SEP-10 web auth tests to the integration suite
Relocate the testanchor.stellar.org authentication tests out of the unit suite into Integration/SEP010Test.php so CI no longer makes live network calls. The tests are unchanged.
